fix(libfdt): typecast operands to match data type
This corrects the MISRA violation C2012-10.3:
The value of an expression shall not be assigned to an object with a
narrower essential type or of a different essential type category.
The operand is explicitly typecasted to the required type
to match the operand/return type.

State: waiting
Link: https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/42598

+++ lib/libfdt/libfdt.h
@@ -1753,7 +1753,7 @@ static inline int fdt_setprop_u32(void *fdt, int nodeoffset, const char *name,
 				  uint32_t val)
 {
 	fdt32_t tmp = cpu_to_fdt32(val);
-	return fdt_setprop(fdt, nodeoffset, name, &tmp, sizeof(tmp));
+	return fdt_setprop(fdt, nodeoffset, name, &tmp, (int)sizeof(tmp));
 }
 
 /**
